TS-40425 Fixed commit being ignored in end commit resolution
