+++ teamscale-maven-plugin/src/main/java/com/teamscale/maven/upload/CoverageUploadMojo.java
@@ -1,5 +1,19 @@
 package com.teamscale.maven.upload;
 
+import com.google.common.base.Strings;
+import com.teamscale.maven.TeamscaleMojoBase;
+import org.apache.maven.plugin.MojoExecutionException;
+import org.apache.maven.plugin.MojoFailureException;
+import org.apache.maven.plugins.annotations.LifecyclePhase;
+import org.apache.maven.plugins.annotations.Mojo;
+import org.apache.maven.plugins.annotations.Parameter;
+import org.apache.maven.plugins.annotations.ResolutionScope;
+import org.apache.maven.project.MavenProject;
+import org.codehaus.plexus.util.xml.Xpp3Dom;
+import shadow.com.teamscale.client.CommitDescriptor;
+import shadow.com.teamscale.client.EReportFormat;
+import shadow.com.teamscale.client.TeamscaleClient;
+
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
@@ -9,21 +23,6 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
-
-import org.apache.maven.plugin.MojoFailureException;
-import org.apache.maven.plugins.annotations.LifecyclePhase;
-import org.apache.maven.plugins.annotations.Mojo;
-import org.apache.maven.plugins.annotations.Parameter;
-import org.apache.maven.plugins.annotations.ResolutionScope;
-import org.apache.maven.project.MavenProject;
-import org.codehaus.plexus.util.xml.Xpp3Dom;
-
-import com.google.common.base.Strings;
-import com.teamscale.maven.TeamscaleMojoBase;
-
-import shadow.com.teamscale.client.CommitDescriptor;
-import shadow.com.teamscale.client.EReportFormat;
-import shadow.com.teamscale.client.TeamscaleClient;
 
 /**
  * Run this goal after the Jacoco report generation to upload them to a
@@ -116,7 +115,8 @@ public class CoverageUploadMojo extends TeamscaleMojoBase {
 	private TeamscaleClient teamscaleClient;
 
 	@Override
-	public void execute() throws MojoFailureException {
+	public void execute() throws MojoFailureException, MojoExecutionException {
+		super.execute();
 		if (skip) {
 			getLog().debug("Skipping since skip is set to true");
 			return;
@@ -127,8 +127,7 @@ public class CoverageUploadMojo extends TeamscaleMojoBase {
 		}
 		teamscaleClient = new TeamscaleClient(teamscaleUrl, username, accessToken, projectId);
 		getLog().debug("Resolving end commit");
-		resolveCommit();
-		resolveRevision();
+		resolveCommitOrRevision();
 		getLog().debug("Parsing Jacoco plugin configurations");
 		parseJacocoConfiguration();
 		try {
